+++ site/blueprints/article.video.php
@@ -4,6 +4,9 @@ title: Video post
 pages: false
 files: true
 fields:
+  articleMeta:
+    label: Article Meta
+    type: headline
   title:
     label: Title
     type:  text
@@ -20,8 +23,9 @@ fields:
   tags:
     label: Tags
     type:  tags
-  line-a:
-    type: line
+  articleContent:
+    label: Article Content
+    type: headline
   text:
     label: Text
     type:  textarea