+++ site/templates/feed.php
@@ -1,12 +1,9 @@
 <?php
 
-$articles = $pages->find('blog')->children()->visible()->flip()->limit(10);
-
-snippet('feed', array(
-  'link'  => url('blog'),
-  'items' => $articles,
-  'description'  => 'text'
+echo page('blog')->children()->visible()->flip()->limit(10)->feed(array(
+  'title'       => $page->title(),
+  'description' => $page->description(),
+  'link'        => 'blog',
 ));
 
 ?>
-